# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API documentation for Project.",
        "title": "Qpay API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/transaction/create/{route}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new transaction with the provided payment details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new transaction for the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gateway route",
                        "name": "route",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction details",
                        "name": "transactionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Bind Error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "403": {
                        "description": "{\"status\": \"error\", \"message\": \"SHEBA doesn't match your credentials\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"No gateway with such route\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in create transaction\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/transaction/filter": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Filter transactions based on start date, end date, and payment amount range.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Filter transactions for the authenticated user",
                "parameters": [
                    {
                        "description": "Filter criteria",
                        "name": "filterRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.FilterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of filtered transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.TransactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Bind Error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/transaction/find/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific transaction associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Find a transaction by ID for the authenticated user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Transaction ID is not correct\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"Transaction does not exist!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/transaction/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all transactions associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List all transactions for the authenticated user",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.TransactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"You don't have any transaction\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/transaction/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Begin a payment transaction with the provided payment details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Begin a payment transaction",
                "parameters": [
                    {
                        "description": "Payment transaction details",
                        "name": "transactionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PaymentTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment transaction details",
                        "schema": {
                            "$ref": "#/definitions/handlers.PaymentTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Invalid transaction details\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "403": {
                        "description": "{\"status\": \"error\", \"message\": \"Transaction is Canceled\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/transaction/start/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get transaction details to start a payment transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction details for starting a transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details for start",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransactionStartResponse"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/transaction/verify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Verify a transaction by tracking code and retrieve its details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Verify a transaction by tracking code",
                "parameters": [
                    {
                        "description": "Verification request",
                        "name": "verifyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Bind Error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"Transaction does not exist!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CreateTransactionRequest": {
            "type": "object",
            "properties": {
                "payment_amount": {
                    "type": "number"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "handlers.FilterRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "max_amount": {
                    "type": "number"
                },
                "min_amount": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "handlers.PaymentTransactionRequest": {
            "type": "object",
            "properties": {
                "card_month": {
                    "type": "integer"
                },
                "card_year": {
                    "type": "integer"
                },
                "cvv2": {
                    "type": "integer"
                },
                "password": {
                    "type": "integer"
                },
                "payment_confirmation": {
                    "description": "دستور پرداخت و کم کردن موجودی (کنسل تراکنش - پرداخت)",
                    "type": "boolean"
                },
                "purchaser_card": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.PaymentTransactionResponse": {
            "type": "object",
            "properties": {
                "payment_amount": {
                    "type": "number"
                },
                "purchaser_card": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tracking_code": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TransactionResponse": {
            "type": "object",
            "properties": {
                "payment_amount": {
                    "type": "number"
                },
                "payment_date": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "purchaser_card": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tracking_code": {
                    "type": "string"
                }
            }
        },
        "handlers.TransactionStartResponse": {
            "type": "object",
            "properties": {
                "owner_name": {
                    "type": "string"
                },
                "payment_amount": {
                    "type": "number"
                }
            }
        },
...
        "handlers.VerifyTransactionRequest": {
            "type": "object",
            "properties": {
                "tracking_code": {
                    "type": "string"
                }
            }
        }
    }
}